setting default SSA to use fixations
Two major changes to the SSA generation:

1). adding options to use fixationStart to drive the gaze cursor, in addition to using gaze samples. Setting fixation as the default.

See issue #104

If we use fixationStart, we don't need to scan video frames, which is a LOT faster. The SSA file is smaller, and fixations look more stable. The down side is that we are only using fixationStart, not fixationEnd, SaccadeStart/End events, meaning that a fixation on the screen may be longer than it actually was. 


2). adding a lot of debugging print, to be controlled by 2 parameters PRINT_VERBOSE and INFO_ONLY. set defaults to False.